refactor: extract remaining section view models
Complete WP-Q3 T5's section split of useDesktopShellViewModels: move the
channels/live/game, profile/author, messages, and settings (five panels)
projections into viewModels/ hooks shaped like the existing
useTimelineViewModels. Logic, dependency arrays, and the facade's return
shape are unchanged (pure moves); shell-wide projections outside the T5
section list (chrome, topic nav, composer, thread, route copy) stay in
the facade, which shrinks 866 -> 450 lines.

Recovers master plan §9.2 B4 (the Q3 T5 scope reduction found by the
2026-07-13 completion review).
